requirements.txt
PyQt5
psycopg2
pillow
qt_material
qdarkstyle
